HDDS-7156. Reset pending delete block count (#4324)
* HDDS-7156: Reset pending delete block count

Reset pending delete block count to 0 if a container has no transactions
Each container has a counter that tracks the number of pending delete
blocks in the delete_txns table for that container. In some cases, this
counter can remain positive when all delete transactions are cleared,
and the counter will never be reset. This means it will continue to be
chosen for deletion instead of other containers that actually have
delete txns.

After this change, if a container is chosen for deletion because it has
>0 pending delete block count but upon deleting from the container, it
is found that there are actually no delete txns, reset the pending
delete block counter to 0. This way it will not be chosen for block
deletion until it gets more delete transactions.

A new unit test for TestBlockDeletingService was added, and the whole
TestBlockDeletingService suite passes.

Author: Ethan Rose
Change-Id: I43e6815962f2db83a95a396020cfa6b7195a8f5f

* Adjust for code changes upstream.

Adjust for renames done in HDDS-7441.

Change-Id: I5e14962e11fa56d959a78ab7f3c99bdd6ab04d18

* Checkstyle

---------

Co-authored-by: Ethan Rose <erose@cloudera.org>
Co-authored-by: Ethan Rose <erose@apache.org>